+++ components/Common/Footer.tsx
@@ -1,8 +1,7 @@
 import { footerData } from '@/data/footer-data'
 import Image from 'next/image'
 import Link from 'next/link'
-import ThemeToggle from '@/components/Common/Theme-Toggle'
-import { FooterWrapper } from '@/components/Common/Footer-Wrapper'
+import { FooterWrapper } from '@/components/Common/FooterWrapper'
 
 export default function Footer() {
   const currentYear = new Date().getFullYear()
@@ -64,8 +63,6 @@ export default function Footer() {
           <div className="mr-4 text-sm text-muted-foreground">
             &copy; {currentYear} {footerData.copyright}
           </div>
-
-          {/*<ThemeToggle />*/}
         </div>
       </FooterWrapper>
     </>
